fix up cross-platform support
