+++ pkg/plugins/action_windows.go
@@ -1,6 +1,1 @@
 package plugins
-
-import "os/exec"
-
-func Setpgid(cmd *exec.Cmd) {
-}
