+++ pkg/plugins/plugin.go
@@ -13,7 +13,6 @@ import (
 	"runtime"
 	"strings"
 	"sync"
-	"text/template"
 	"time"
 
 	"github.com/pkg/errors"
@@ -281,40 +280,6 @@ func (p *Plugin) CurrentCycleItem() *Item {
 		p.CycleIndex = 0
 	}
 	return p.Items.CycleItems[p.CycleIndex]
-}
-
-func (p *Plugin) runInTerminal(appleScriptTemplate3, command, paramsStr string, vars []string) error {
-	tpl, err := template.New("appleScriptTemplate3").Parse(appleScriptTemplate3)
-	if err != nil {
-		return err
-	}
-	commandLine := command
-	var renderedScript bytes.Buffer
-	err = tpl.Execute(&renderedScript, struct {
-		Command string
-		Vars    string
-		Params  string
-	}{
-		Command: commandLine,
-		Vars:    fmt.Sprintf("%q", variablesEnvString(vars)),
-		Params:  paramsStr,
-	})
-	if err != nil {
-		return err
-	}
-	appleScript := renderedScript.String()
-	log.Println(p.Command, "RunInTerminal", appleScript)
-	cmd := exec.Command("osascript", "-s", "h", "-e", appleScript)
-	var stderr bytes.Buffer
-	cmd.Stderr = &stderr
-	err = cmd.Run()
-	if err != nil {
-		p.Debugf("(ignoring) RunInTerminal failed: %s", err)
-	}
-	if cmd.ProcessState != nil && cmd.ProcessState.ExitCode() != 0 {
-		return errors.Errorf("run in terminal script failed: %s", stderr.String())
-	}
-	return nil
 }
 
 // RunInTerminal runs this plugin in a terminal using the template
