--- pkg/update/update_unix.go
@@ -1,13 +1,0 @@
-//go:build linux || darwin
-package update
-
-import (
-	"os/exec"
-	"syscall"
-)
-
-func Setpgid(cmd *exec.Cmd) {
-	cmd.SysProcAttr = &syscall.SysProcAttr{
-		Setpgid: true,
-	}
-}
