--- pkg/update/update_windows.go
@@ -1,6 +1,0 @@
-package update
-
-import "os/exec"
-
-func Setpgid(cmd *exec.Cmd) {
-}
